fix(experiments): drop a search example the scoped registry cannot resolve
The placeholder examples are authored per view, but the registry is now built
per table instance, so a dataset-scoped page dropped the Dataset facet while
still advertising `dataset:legal-answer-quality` as a hint for a field it
rejects. Keep the list authored and only filter it: an example whose field the
registry cannot resolve is not offered. Aliases and `scores.` paths resolve
through the registry, so the check costs one throwaway build and the unscoped
placeholder is unchanged.

The view-specific registry tests move next to the registry, where the filter
config they need is a same-feature import.
